Revert "New header, brand and theme colors"
