--- .dockerignore
@@ -1,7 +1,0 @@
-.venv
-environment/*
-db.sqlite3
-
-# Assets
-assets/
-static/
