+++ assets/css/drag_and_drop.css
@@ -1,86 +1,12 @@
-.btn-file {
-    position: relative;
-    overflow: hidden;
-}
-
-.btn-file input[type="file"] {
-    position: absolute;
-    top: 0;
-    right: 0;
-    min-width: 100%;
-    min-height: 100%;
-    filter: alpha(opacity=0);
-    opacity: 0;
-    outline: none;
-    cursor: inherit;
-    display: block;
-}
-
-.btn-file.btn-image {
-    height: fit-content;
-    display: block;
-    width: 100%;
-}
-
-
-.btn-file.btn-image>img {
-    min-width: 100%;
-    width: 100%;
-    min-height: 100%;
-    height: 100%;
-}
-.btn-file.btn-image>img.square-image {
-    object-fit: cover;
-    object-position: center;
-    aspect-ratio: 1;
-}
-
-
-
-.dropzone {
-    border: 2px dashed rgba(0, 99, 255, 1);
+.dropzone{
+    border: 2px dashed rgba(0,99,255,1);
     border-radius: 2px;
     width: 100%;
     height: 100%;
-    aspect-ratio: 1;
     min-height: 200px;
     display: flex;
     align-items: center;
     justify-content: center;
     content: "Arrastra aqui una imagen.";
     color: black;
-    /* opacity: 50%; */
-    padding: 0.5rem;
-}
-
-.dropzone>img {
-    min-width: 100%;
-    width: 100%;
-    min-height: 100%;
-    height: 100%;
-    object-fit: cover;
-    object-position: center;
-    aspect-ratio: 1;
-}
-
-.dropzone .content-text {
-    position: absolute;
-    color: white;
-    font-weight: 700;
-}
-
-.btn-file input[type="file"]~.dropzone {
-    display: flex;
-}
-
-.btn-file input[type="file"].fill~.dropzone {
-    display: none;
-}
-
-.btn-file input[type="file"]~.image-preview {
-    display: none;
-}
-
-.btn-file input[type="file"].fill~.image-preview {
-    display: block;
 }
