--- assets/css/footer.css
@@ -1,70 +1,0 @@
-/* -------------- FOOTER --------------- */
-.footer {
-    /* position: fixed; */
-    bottom: 0;
-    /* width: 100vw; */
-    box-sizing: border-box;
-    margin-top: 30px;
-    padding: 8px 80px;
-    height: auto;
-    display: flex;
-    justify-content: space-evenly;
-    background-color: #5e0f35;
-    color: #ffffff;
-    box-shadow: 0px -2px 20px 0px #28332bca;
-}
-
-.footer ul {
-    list-style: none;
-    padding: 0px;
-}
-
-.footer .footer-content {
-    width: inherit;
-    display: flex;
-    justify-content: space-evenly;
-}
-
-.footer a {
-    text-decoration: none;
-    color: #ffffff;
-}
-
-.footer a.creator {
-    color: #d6149f;
-    font-size: 14px;
-}
-
-.footer p {
-    margin: 0px;
-    margin-top: 10px;
-    font-size: 12px;
-}
-
-.footer h3 {
-    margin-top: 0px;
-}
-
-.footer li {
-    margin-bottom: 5px;
-    display: flex;
-    justify-content: center;
-}
-
-.footer .redes {
-    box-sizing: border-box;
-    margin-top: 10px;
-}
-
-.footer .redes h4 {
-    margin: 0px;
-}
-
-.footer .redes ul {
-    display: flex;
-    justify-content: center;
-}
-
-.footer .redes ul img {
-    width: 60px;
-}
