--- assets/css/header.css
@@ -1,116 +1,0 @@
-/* -------------- HEADER ------------- */
-.header {
-    position: fixed;
-    top: 0;
-    padding: 0px 150px;
-    height: 80px;
-    width: 100vw;
-    z-index: 5;
-    box-sizing: border-box;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    background: rgb(24, 23, 31);
-    background: linear-gradient(45deg, var(--stari-main-gradient-color-1) 55%, var(--stari-main-gradient-color-2) 100%);
-    box-shadow: 0px 2px 20px 0px #98bcff80;
-}
-
-.header-padding {
-    height: 80px;
-    margin-bottom: 20px;
-}
-
-/* Header-> Profile picture */
-.header .profile-picture-container {
-    height: 60px;
-    width: 60px;
-}
-
-.header .profile-picture img {
-    box-shadow: 5px 5px 5px 0px #01093391;
-}
-
-.header .profile-picture-container {
-    overflow: initial;
-}
-
-.brand {
-    display: flex;
-    justify-content: space-evenly;
-    align-items: center;
-    letter-spacing: 2px;
-}
-
-.brand .brand-element {
-    margin: 0px 10px;
-}
-
-.brand .image {
-    height: 50px;
-    width: auto;
-    background-color: #ffffff;
-    margin: 0px 15px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    border-radius: 40px;
-    box-shadow: 0px 3px 15px 2px #98bcff80;
-}
-
-.brand .image img {
-    height: 40px;
-    width: 40px;
-    margin: 0px 10px;
-}
-
-.brand h1 {
-    font-size: 30px;
-    font-weight: lighter;
-    margin: 0px;
-    color: #ffffff;
-    text-shadow: 0px 5px 10px #cccccc4d;
-}
-
-.navbar ul {
-    list-style: none;
-    padding: 0px 30px;
-    margin: 0px;
-    height: 60px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-family: 'Thasadith', sans-serif;
-}
-
-.navbar li {
-    margin: 0px;
-    height: inherit;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding: 0px 10px;
-}
-
-.navbar li:hover {
-    background-color: #002a77;
-    box-sizing: border-box;
-    border-radius: 2px 2px 20px 20px;
-    box-shadow: 0px 3px 15px 2px #395f9c80;
-}
-
-.navbar li a {
-    text-decoration: none;
-    color: #ffffff;
-    font-size: 15px;
-    font-weight: lighter !important;
-}
-
-.navbar ul a {
-    margin: 0px 15px;
-}
-
-.navbar ul img.icon {
-    height: 20px;
-    width: 20px;
-    filter: invert(100%);
-}
