--- assets/css/posts/new.css
@@ -1,4 +1,0 @@
-/* .posts-grid {
-    display: flex;
-    justify-content: center;
-} */
