--- assets/css/posts/posts-grid.css
@@ -1,15 +1,0 @@
-.posts-grid {
-    margin-top: -0.5rem;
-    position: relative;
-}
-
-.post-container {
-    width: 50%;
-    position: absolute;
-}
-
-@media (max-width: 568px) {
-    .post-container {
-        width: 100%;
-    }
-}
