--- assets/css/posts/posts.css
@@ -1,66 +1,0 @@
-.post {
-    margin: 0.5rem;
-    padding: 0px;
-    box-sizing: border-box;
-    box-shadow: 0px 0px 6px -5px var(--stari-secondary-shadow-color);
-    border-radius: 10px;
-    background-color: var(--stari-secondary-bg-color);
-    color: var(--stari-post-text);
-    transition: all 0.1s ease;
-    position: relative;
-    height: fit-content;
-}
-
-.post .post-url {
-    color: var(--stari-post-text);
-    position: absolute;
-    left: 0;
-    top: 0;
-    bottom: 0;
-    right: 0;
-}
-
-.post .post-url:hover {
-    color: var(--stari-post-text);
-    text-decoration: none;
-}
-
-/* .post .content-post {
-    border-radius: 10px;
-} */
-
-.post:has(.post-url:hover) {
-    background-color: var(--stari-tertiary-bg-color);
-    transform: scale(1.005);
-    filter: brightness(0.95);
-}
-
-.post-title {
-    font-size: 1.75rem;
-    font-weight: 700;
-}
-
-.post-photo {
-    width: 100%;
-}
-.post-data{
-    color: var(--stari-post-text-1);
-    display: grid;
-    gap: 0rem;
-    /* align-items: center; */
-}
-.post-header:hover{
-    text-decoration: none;
-}
-.post-user{
-    margin: 0;
-    font-size: 1rem;
-}
-.post-date{
-    font-size: 0.75rem;
-    color: var(--stari-post-text-2);
-}
-.post-user-picture {
-    width: 2.5rem;
-    height: 2.5rem;
-}
