+++ assets/css/sesion.css
@@ -27,6 +27,3 @@ body{
 .invalid-feedback{
     text-align: start!important;
 }
-.btn {
-    border-radius: 50px 50px 50px 50px;
-}
