+++ pyproject.toml
@@ -14,10 +14,6 @@ Django = "^5.0.7"
 pillow = "^10.4.0"
 pytz = "^2024.1"
 sqlparse = "^0.5.1"
-daphne = "^4.1.2"
-django-storages = {extras = ["s3"], version = "^1.14.3"}
-twisted = {extras = ["http2", "tls"], version = "^24.3.0"}
-mysqlclient = "^2.2.4"
 
 
 [build-system]
