--- static/css/test_prod.css
@@ -1,3 +1,0 @@
-/* This file is empty, it is just for test if the deploy has the 
- * the assets resulting of the collectstatic ran with github actions
- */
