--- tests/testfile.txt
@@ -1,1 +1,0 @@
-This is a testfile hehehe
